refactor: 요금 정책 인터페이스 제거
